Move audit_log utility function to openedx/core
